requirements.txt
PyYAML
docker
GitPython
